+++ addon/components/power-select-multiple.js
@@ -1,0 +1,29 @@
+import Ember from 'ember';
+import layout from '../templates/components/power-select-multiple';
+
+export default Ember.Component.extend({
+  layout,
+  selectedComponent: 'power-select-multiple/selected',
+  beforeOptionsComponent: null,
+  optionsComponent: 'power-select-multiple/options',
+
+  // Actions
+  actions: {
+    handleOpen(select, e) {
+      let action = this.get('onopen');
+      if (action) { action(select, e); }
+      this.focusInput();
+    },
+
+    handleFocus(select, e) {
+      let action = this.get('onfocus');
+      if (action) { action(select, e); }
+      this.focusInput();
+    }
+  },
+
+  // Methods
+  focusInput() {
+    this.element.querySelector('.ember-power-select-trigger-multiple-input').focus();
+  }
+});