+++ addon/components/power-select/single/selected.js
@@ -1,7 +1,0 @@
-import Ember from 'ember';
-import layout from '../../../templates/components/power-select/single/selected';
-
-export default Ember.Component.extend({
-  layout: layout,
-  tagName: ''
-});